Add constraints to the database tables
In most tables the FOREIGN KEY is the employee_key or the branch_id.

Small parts of the German documentation have been adapted.
The English documentation is now behind and out of sync with the German original.

+++ src/sql/principle_roster_archive.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS `principle_roster_archive` (
   `primary_key` int(10) unsigned NOT NULL AUTO_INCREMENT,
   `alternating_week_id` tinyint(4) NOT NULL,
-  `employee_key` int(10) NOT NULL,
+  `employee_key` int(10) unsigned NOT NULL,
   `weekday` tinyint(4) NOT NULL,
   `duty_start` time NOT NULL,
   `duty_end` time NOT NULL,
@@ -9,7 +9,7 @@ CREATE TABLE IF NOT EXISTS `principle_roster_archive` (
   `break_end` time DEFAULT NULL,
   `comment` text COLLATE utf8mb4_unicode_ci DEFAULT NULL,
   `working_hours` float DEFAULT NULL,
-  `branch_id` int(11) NOT NULL DEFAULT 1,
+  `branch_id` tinyint(3) unsigned NOT NULL DEFAULT 1,
   `was_valid_until` date DEFAULT current_timestamp(),
   PRIMARY KEY (`primary_key`)
 ) ENGINE=InnoDB  DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci
